feat: Disable Service Links by default
In the default configuration, Kubernetes will inject a bunch of
environment variables for each service in the Namespace. The idea is to
aid with service discovery, but while they are rarely used in practice,
the injected environment variables might interfere that try to determine
their configuration from environment variables.

This commit disables this behaviour by default by setting
`enableServiceLinks: false` in the pod specs, but allows users to
reenable the links using a label.

Signed-off-by: Manuel Hutter <manuel@hutter.io>